exportしているtypeを.d.tsなどにまとめた
