+++ src/@types/ApiAccess.d.ts
@@ -1,0 +1,7 @@
+export type EndPointCOnfig = {
+  endPointUrl: string;
+};
+
+export type Config = {
+  config: EndPointCOnfig;
+};
