+++ src/@types/Registration.d.ts
@@ -1,0 +1,4 @@
+export type reloadState = {
+  isReload: boolean;
+  caller: string;
+};
